+++ websitebuild-20250429/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-06-18 Version: 2.28.1
+- Update API ListAppInstances: add request parameters BizIds.
+- Update API NotifyAppNotificationForAdmin: add request parameters Payload.
+
+
 2026-06-16 Version: 2.28.0
 - Support API CheckAppVerifyCode.
 - Support API CreateAppNotificationScene.